@@ -1,6 +1,6 @@
 module github.com/smirl/highheath
 
-go 1.17
+go 1.19
 
 require (
 	github.com/bradleyfalzon/ghinstallation/v2 v2.1.0
